readded f.seek(header["cinefileheader"].OffSetup)
in fact I don't know, why I don't had it in my version. It seems to work like this too. 
In the file docu (version 2020) there is no gap between the BITMAPINFOHEADER and the SETUP part, so it should be fine without the seek(). But for savety reasons this can be left in.